+++ helpers/index.js
@@ -1,3 +1,7 @@
 module.exports = {
   cleanup: require('./cleanup'),
+  rollDice: require('./rollDice'),
+  explode: require('./explode'),
+  standard: require('./standard'),
+  generate: require('./generate'),
 }
