+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "swadebot",
-  "version": "1.8.0",
+  "version": "1.8.2",
   "description": "",
   "main": "index.js",
   "scripts": {
